Remove unused d3 library, add build folder to kubernetes example by changing git ignore
